1.引入ktor作为部分请求框架。 2.将更多异步请求改为协程请求。 3.优化播放器，新增加载播放动画。 1. Introduce ktor as a partial request framework. 2. Change more asynchronous requests to coroutine requests. 3. Optimize the player and add loading and playing animations.
Former-commit-id: 7701792348d716c9cf2a0e8f66cf04ec22c2f70f

+++ app/src/main/java/com/imcys/bilibilias/home/ui/model/BangumiSeasonBean.java
@@ -1079,7 +1079,7 @@ public class BangumiSeasonBean implements Serializable {
             private String avatar_subscript_url;
             private int follower;
             private int is_follow;
-            private int mid;
+            private long mid;
             private String nickname_color;
             private PendantBean pendant;
             private int theme_type;
@@ -1121,11 +1121,11 @@ public class BangumiSeasonBean implements Serializable {
                 this.is_follow = is_follow;
             }
 
-            public int getMid() {
+            public long getMid() {
                 return mid;
             }
 
-            public void setMid(int mid) {
+            public void setMid(long mid) {
                 this.mid = mid;
             }
